Create Depth_First-search.py
the code above, we define two functions: recursive_dfs and iterative_dfs, to perform a Depth-First Search on a given graph. You can call these functions by providing the graph represented as an adjacency list and the starting node for the traversal.

The code will print the nodes in the order they are visited during the DFS traversal. You can change the starting node by passing a different node to the functions.